# public/api-specs.yaml
openapi: 3.0.0
info:
  title: CoderManagement
  version: "1.0.0"
paths:
...
  /tasks/{id}:
    get:
      tags:
        - Task
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: id of the task
      responses:
        '400':
          description: Bad request
        '404':
          description: Task not found
        '200':
          description: Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    put:
      tags:
        - Task
      summary: Update an existing task
      operationId: updateTask
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: id of the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                assignee:
                  type: string
                removeAssignee:
                  type: string
                  enum: [no, yes]
      responses:
        '400':
          description: Invalid request body
        '404':
          description: Task not found
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      tags:
        - Task
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: id of the task
      responses:
        '400':
          description: Invalid request body
        '404':
          description: Task not found
        '204':
          description: Deleted, no content
...
components:
  schemas:
...
    Task: 
      type: object
      properties:
        id:
          type: string
          format: mongo_id
        name: 
          type: string
        description:
          type: string
        status: 
          type: string
          description: default to pending if empty
          enum: [pending, working, review, done, archive]
        assignee:
          type: object
          properties:
            id:
              type: string
              format: mongo_id
            name:
              type: string
            role:
              type: string
        is_deleted: 
          type: boolean
          example: false
        updated_at: 
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
      required: 
        - name
        - description
...